+++ CHANGES.rst
@@ -14,8 +14,11 @@ Version 8.1.4
 
 Unreleased
 
+-   Replace all ``typing.Dict`` occurrences to ``typing.MutableMapping`` for
+    parameter hints.  :issue:`2255`
 -   Improve type hinting for decorators and give all generic types parameters.
     :issue:`2398`
+-   Fix return value and type signature of `shell_completion.add_completion_class` function. :pr:`2421`
 
 
 Version 8.1.3